+++ frontend/src/types/user/addRole.dto.ts
@@ -1,0 +1,5 @@
+
+export interface AddRoleDto {
+	userId: number;
+	value: string;
+}