+++ frontend/src/types/user/createUser.dto.ts
@@ -1,0 +1,6 @@
+export interface CreateUserDto {
+	email: string;
+	username: string;
+	fullname: string;
+	password: string;
+}